chore: clean doc, remove dead code, format sources (#1490)
