+++ OUDS/Core/Components/Sources/_/Extensions/View+TabBar.swift
@@ -13,6 +13,62 @@
 
 import SwiftUI
 
+#if os(iOS)
+
+extension View {
+
+    /// Get the tab bar within the view hierarchy
+    public func findTabBar() -> UITabBar? {
+        guard let windowScene = UIApplication.shared.connectedScenes.first as? UIWindowScene,
+              let window = windowScene.windows.first else { return nil }
+
+        // Find via root view controller
+        if let tabBar = findTabBar(in: window.rootViewController) {
+            return tabBar
+        }
+
+        // Or otherwise check in views
+        return findTabBar(in: window)
+    }
+
+    /// Recursive search in view controller to find the tab bar
+    private func findTabBar(in viewController: UIViewController?) -> UITabBar? {
+        guard let viewController else { return nil }
+
+        if let tabBarController = viewController as? UITabBarController {
+            return tabBarController.tabBar
+        }
+
+        for child in viewController.children {
+            if let tabBar = findTabBar(in: child) {
+                return tabBar
+            }
+        }
+
+        if let presented = viewController.presentedViewController {
+            return findTabBar(in: presented)
+        }
+
+        return nil
+    }
+
+    /// Recursive search in view to find the tab bar
+    private func findTabBar(in view: UIView) -> UITabBar? {
+        if let tabBar = view as? UITabBar {
+            return tabBar
+        }
+
+        for subview in view.subviews {
+            if let tabBar = findTabBar(in: subview) {
+                return tabBar
+            }
+        }
+
+        return nil
+    }
+}
+#endif
+
 // MARK: - PreferenceKey
 
 /// A SwiftUI `PreferenceKey` that propagates tab bar visibility intent **upward** through the
@@ -31,9 +87,11 @@ import SwiftUI
 /// The `reduce` function uses logical OR so that if any descendant hides the tab bar, the
 /// preference is `true`. When all descendants that hid the bar are removed from the hierarchy,
 /// SwiftUI resets the preference to the `defaultValue` (`false`), making the overlays reappear.
-public struct TabBarHiddenPreferenceKey: PreferenceKey {
-    nonisolated(unsafe) public static var defaultValue: Bool = false
-    public static func reduce(value: inout Bool, nextValue: () -> Bool) {
+struct TabBarHiddenPreferenceKey: PreferenceKey {
+
+    nonisolated(unsafe) static var defaultValue: Bool = false
+
+    static func reduce(value: inout Bool, nextValue: () -> Bool) {
         value = value || nextValue()
     }
 }
@@ -45,24 +103,31 @@ extension View {
     /// Hides the native tab bar **and** signals `OUDSTabBar` to hide its overlay elements
     /// (selected tab indicator and top divider) in sync.
     ///
-    /// Use this modifier instead of — or in addition to — `.toolbar(.hidden, for: .tabBar)`
+    /// Use this modifier instead of `.toolbar(.hidden, for: .tabBar)`
     /// whenever a view inside a `NavigationStack` embedded in an `OUDSTabBar` needs to hide
     /// the tab bar. Without this modifier, `OUDSTabBar` cannot detect the visibility change
-    /// and its overlay elements remain visible on top of the content.
+    /// and its overlay elements remain visible on top of the content. This helper must be used for
+    /// apps with nested views and tab bars.
+    ///
+    /// Otherwise for simple app `.toolbar(.hidden, for: .tabBar)` is enough.
     ///
     /// ```swift
     /// struct DetailView: View {
     ///     var body: some View {
     ///         Text("Detail")
-    ///             .oudsHideTabBar()  // hides UITabBar + OUDS overlays
+    ///             .oudsHideTabBar()
     ///     }
     /// }
     /// ```
     ///
-    /// - Returns: A view that hides the tab bar and propagates the hidden state to `OUDSTabBar`.
+    /// - Returns: A view that hides the tab bar and propagates the hidden state to `OUDSTabBar` overlay items.
     @available(iOS 16, *)
     public func oudsHideTabBar() -> some View {
+        #if !os(macOS)
         toolbar(.hidden, for: .tabBar)
             .preference(key: TabBarHiddenPreferenceKey.self, value: true)
+        #else
+        preference(key: TabBarHiddenPreferenceKey.self, value: true)
+        #endif
     }
 }
